ADD(artifacts): Configuration Classes
Adding AppContext And AppSetup Classes

AppSetup responsible for holding all helper functions needed to help get
the application in correct state when starting. Need this class since no
configuration is built in, since the app is not using Maven or something
like Spring framework right now.

Moreover, AppContext is used to hold the state of the CLI aplication. It
is concerned with keeping track of the active customer and active CLI menu.